requirements.txt
faiss-cpu
numpy
sentence-transformers
torch
google-generativeai
openai
ragas
litellm
groq